Deprecate X-Android-{Transports,Protocols} headers for configuration.
This feature was introduced in 4910039a572abc6d20cac48eb2c5e92591520537
and documented in the changelog for version 1.1. The synthetic header
'X-Android-Selected-Transport', which is used for querying, was renamed
 - to OkHttp-Selected-Transport in OkHttp 1.2, and then
 - to OkHttp-Selected-Protocol in OkHttp 1.5.
It is still supported.

After this CL, the two headers no longer change OkHttpURLConnection's
configuration. They continue to not be sent to the server.

The benefit of deprecation is that we don't carry an obscure, untested
feature which may not work as planned.

The deprecated feature, as well as the fact that these headers were not
sent to the server, turned out to be untested.