HeldCertificate: Don't harc-code BC provider.
This test code used by MockWebServer hard-codes the provider
(BC = BouncyCastle) for RSA crypto; since the algorithm is
tightly specified, there should not be any reason to specify
any particular provider; not all platforms may include the
BC provider.

This change changes the code to not specify a provider,
which allows the platform to pick an appropriate provider
for the specified algorithm.